+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal 0.12.0.9011
+# teal 0.12.0.9012
 
 ### Major breaking changes